improve oracle compatibility
For whatever reason, Atlassians AO library replaces Order("item.ID")
where item is alias for LabelItem.class's table to "item"."ID" which
obvsiously causes a SQL exception because Oracle treats "item" as table,
but not as alias.

This changeset needs to be checked on other databases including h2 and
postgre.